package/requirements.txt
absl-py==0.13.0
matplotlib==3.3.4
pandas==1.2.5
scikit-learn
tensorflow
torch==1.9.0
transformers==4.8.2
